refactored the long to int

+++ ReviewService/src/main/java/io/app/service/ReviewService.java
@@ -6,6 +6,6 @@ import org.springframework.data.domain.Page;
 
 public interface ReviewService {
     public ApiResponse createReview(Review review);
-    public Page<Review> getReviewsByCompanyId(long companyId,long pageNo,long size);
-    public Page<Review> getReviewsByJobId(long jobId,long pageNo,long size);
+    public Page<Review> getReviewsByCompanyId(long companyId,int pageNo,int size);
+    public Page<Review> getReviewsByJobId(long jobId,int pageNo,int size);
 }
